src/requirements.txt
dj-database-url==0.4.2
Django==1.11.9
django-cors-headers==2.1.0
django-post-office==3.0.4
djangorestframework==3.7.7
djangorestframework-jwt==1.11.0
gunicorn==19.7.1
jsonfield==2.0.2
Pillow==5.0.0
psycopg2==2.7.3.2
PyJWT==1.5.3
pytz==2017.3
